pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.0",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-jose>=3.5.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "statsmodels>=0.14.4",
    "streamlit-searchbox>=0.1.22",
    "streamlit>=1.46.1",
]
